+++ src/main/java/tpVol/dao/DaoVol.java
@@ -2,6 +2,6 @@ package tpVol.dao;
 
 import tpVol.model.Vol;
 
-public interface DaoVol extends DaoGeneric<Vol, Integer> {
+public interface DaoVol extends DaoGeneric<Vol, Long> {
 
 }
